improve!: remove `always_allow` for now
I originally had both an `always_allow` and an `always_forbid`, but
I don't think either should be necessary. It's likely that using either
should be considered an anti-pattern, as they define a sort of universal
override that is difficult to understand or compose. Instead of relying
on `always_allow`, it should be possible to structure policy definitions
to avoid the `forbid` that you're trying to override. If you don't
include the `forbid` in the first place, there is no need to override
with `always_allow`.